Added title text for user-friendliness
Added title text to the input options to make it a little more clear to users what the different settings will do. Added placeholder text to the Separator input so the user can distinguish between " " (space) separator and "" (default - tab) separator. Field will now show "<tab>" if the default is in use, and " " when space is in use.