Auto-create table when parameters change
Instead of requiring the user to click "Create Table" any time a parameter is changed, the table is automatically populated every time a new parameter is selected, or the input text changes. This way, the user can quickly flip through table styles to see which one they like best, without having to click back and forth.