Reorganized for readability, +wikimedia format
Reorganized the i/j loops so they are hopefully a bit easier to follow, removed some case-specific conditions and used flags to make it more generic. Added several table formats, including wikimedia. The wikimedia format has some extra carriage returns, but it still works in wikipedia, so I plan to leave it as it is.