Parse of tables with arbitrary vertical separators
Identifies vertical separators by recursing through the array to find repeated characters. Does not yet handle spaces as a vertical separator.
Allows user-input character for the output separator (so creating csvs is possible, just put a comma in the Custom Separator box before pasting in the text to parse)
Added error messages to describe parsing failures.
Returns original input (sans separator lines) if a parsing error occurs, so it is less likely that the output will be overwritten automatically if the user changes a setting accidentally.
Corrected typo in isSeparatorLine function name.