Add reStructuredText Grid and reddit formats
ReStructuredText Grid format uses a different separator for the top and bottom of the header, the assumption has been that the header top and bottom are the same when both are present. A more robust solution could be to define headerTop, headerBottom, and bodyBottom horizontal character definitions for each style.

Reddit has no left or right side lines, but has vertical separators in the middle, which was also new. A more robust solution could be to define headerLeft, headerRight, bodyLeft, and bodyRight vertical character definitions for each style.